Moved XeGatherCoalescePass to separate file
